Add 'Given I sleep for <x> seconds' command
This function was modified from the `ecukes` repo,
https://github.com/ecukes/ecukes/blob/master/features/async.feature

It sleeps for the given time in seconds, e.g.

    Given I sleep for 1 second

This is necessary to test `Man-mode`. `Man-mode` initially creates an
empty buffer like `*Man grep*`, then fills it in when the `man` process
completes. An empty buffer causes other tests to fail.

It includes a workaround to `sleep-for` returning early, which is
something I encountered in testing.