add shell-mode support
Save and restore buffers in `shell-mode`, the major mode created by
`M-x shell`. `shell-mode` is similar to `term-mode`, but runs a
`comint-mode` unix shell instead of a full-blown terminal emulator.

The current directory is preserved between boots. If the current
directory no longer exists, it falls back on `/` to avoid errors.

The contents of the buffer are not preserved. I plan to add this in a
future commit.